test: move RPC support classes into a separate package
